point main to cjs file
"main" is a fallback for "exports" now. When a package has an
"exports" field, that takes precedence over "main". But "main" is used
by tools that don't support "exports"—like browserify. So it should
point to the CommonJS file. I've confirmed that this fixes the issue
with bundling this package using browserify.